new file:   Makefile 	new file:   main.cpp 	new file:   rank.cpp 	new file:   rank.h 	new file:   ranker 	new file:   ranksystem.cpp 	new file:   ranksystem.h
Initial commit.

Very much unfinished current development is on the rank system.  I intend to refactor some of the class in rank.cpp to a base class so that ranksystem can access the data pulled without creating a full instance of the main rank class (only done if a ranksystem.conf or some such isn't created.  The whole ranking and searching engine has yet to be written but most of the command line argument processing is there.