modified:   Makefile 	modified:   main.cpp 	modified:   rank.cpp 	modified:   rank.hpp 	modified:   ranksystem.cpp 	modified:   ranksystem.hpp 	renamed:    rank.cpp -> rankutils.cpp 	new file:   rankutils.hpp
Made RankUtils class as common base for rank classes.

Moved former member functions that are shared between the rank classes to a utility base class to inherit from.  Also getRankBook(), which was moved to the utility class, now returns a reference to a const vector <Data> both for performance and convenience.  It is mostly untested at this point, but it compiles.