Completed Ranking Functionality
The program will now find people who need to be ranked, and what rank they need to be given.  Minimal testing has been done with it, and it appears to work.  To see output uncomment lines 64-66 in rank.cpp.  No code has been done to really do output yet, but current and destination rank names and abbreviations are kept in a vector of the new struct PersonToRank, along with the person's name and rankline (all used for processing the ranks).  Output code should be trivial to implement, and will complete the basic ranking functionality.